refactor: refillstock consumer 분리
